Refactor dashboard child into modular components
